Adding modularity to code, coderefactor
