+++ src/lsp/methods/suggestEnvironmentVars.ts
@@ -2,9 +2,13 @@ import * as vscode from "vscode";
 import { ChildProcess, execSync } from "child_process";
 import * as path from "path";
 import * as fs from "fs";
-import triggers from "./triggers";
-import { ErrorCodes, IJSONRPCResponse, logToChannel } from "./lsp/utils";
-import { getEnvsFromLsp } from "./lsp/methods/lspSuggestEnvs";
+import triggers from "../../triggers";
+import {
+  ErrorCodes,
+  IJSONRPCResponse,
+  logToChannel,
+} from "../response/generalResponse";
+import { getEnvsFromLsp } from "../request/suggestEnvVarsRequest";
 
 let envVars = [] as string[];
 let cursorPosition = 0;
@@ -164,20 +168,30 @@ function handleErrorsForSuggestEnvs(response: IJSONRPCResponse) {
   if (response.error) {
     switch (response.error.code) {
       case ErrorCodes.ERR_INVALID_REQUEST:
-        vscode.window.showErrorMessage("Invalid JSON-RPC request: " + response.error.message);
+        vscode.window.showErrorMessage(
+          "Invalid JSON-RPC request: " + response.error.message
+        );
         break;
       case ErrorCodes.ERR_UNSUPPORTED_FEATURE:
-        vscode.window.showErrorMessage(response.error.message, 'Visit GitHub').then(selected => {
-          if (selected === 'Visit GitHub') {
-            vscode.env.openExternal(vscode.Uri.parse('https://github.com/HexmosTech/Lama2'));
-          }
-        });
+        vscode.window
+          .showErrorMessage(response.error.message, "Visit GitHub")
+          .then((selected) => {
+            if (selected === "Visit GitHub") {
+              vscode.env.openExternal(
+                vscode.Uri.parse("https://github.com/HexmosTech/Lama2")
+              );
+            }
+          });
         break;
       case ErrorCodes.ERR_INVALID_AFTER_SHUTDOWN:
-        vscode.window.showErrorMessage("Invalid request after shutdown: " + response.error.message);
+        vscode.window.showErrorMessage(
+          "Invalid request after shutdown: " + response.error.message
+        );
         break;
       default:
-        vscode.window.showErrorMessage("Unknown error: " + response.error.message);
+        vscode.window.showErrorMessage(
+          "Unknown error: " + response.error.message
+        );
     }
   }
 }
